requirements.txt
Flask==2.3.2
redis==4.6.0
python-dotenv==1.0.0
pandas==1.5.3
matplotlib==3.6.2
numpy==1.24.3